Separate all output POJOs into separate classes.
