+++ src/main/java/bio/terra/testrunner/runner/TestRunner.java
@@ -16,8 +16,6 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.*;
 import java.util.concurrent.*;
 import java.util.stream.Collectors;
@@ -34,6 +32,7 @@ public class TestRunner {
   private ThreadPoolExecutor disruptionThreadPool;
   private List<List<Future<UserJourneyResult>>> userJourneyFutureLists;
 
+  // test run outputs
   private List<TestScriptResult> testScriptResults;
   protected TestRunSummary summary;
 
@@ -43,62 +42,6 @@ public class TestRunner {
       new HashMap<String, Map<String, String>>();
 
   private boolean exceptionThrownInCleanup = false;
-
-  public static class TestRunSummary {
-    public String id;
-
-    public long startTime = -1;
-    public long startUserJourneyTime = -1;
-    public long endUserJourneyTime = -1;
-    public long endTime = -1;
-    public List<TestScriptResult.TestScriptResultSummary> testScriptResultSummaries;
-
-    public TestRunSummary() {}
-
-    public TestRunSummary(String id) {
-      this.id = id;
-    }
-
-    private String startTimestamp;
-    private String startUserJourneyTimestamp;
-    private String endUserJourneyTimestamp;
-    private String endTimestamp;
-
-    // Include user-provided TestSuite name in the summary:
-    // This can be used to facilitate grouping of test runner results on the dashboard.
-    private String testSuiteName;
-
-    public String getStartTimestamp() {
-      return millisecondsToTimestampString(startTime);
-    }
-
-    public String getStartUserJourneyTimestamp() {
-      return millisecondsToTimestampString(startUserJourneyTime);
-    }
-
-    public String getEndUserJourneyTimestamp() {
-      return millisecondsToTimestampString(endUserJourneyTime);
-    }
-
-    public String getEndTimestamp() {
-      return millisecondsToTimestampString(endTime);
-    }
-
-    public String getTestSuiteName() {
-      return testSuiteName;
-    }
-
-    public void setTestSuiteName(String testSuiteName) {
-      this.testSuiteName = testSuiteName;
-    }
-
-    private static String millisecondsToTimestampString(long milliseconds) {
-      DateFormat dateFormat =
-          new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSSSSS'Z'"); // Quoted Z to indicate UTC
-      dateFormat.setTimeZone(TimeZone.getTimeZone("UTC"));
-      return dateFormat.format(new Date(milliseconds));
-    }
-  }
 
   protected TestRunner(TestConfiguration config) {
     this.config = config;
@@ -573,9 +516,9 @@ public class TestRunner {
    * Read in the test run summary from the output directory and return the TestRunner.TestRunSummary
    * Java object.
    */
-  public static TestRunner.TestRunSummary getTestRunSummary(Path outputDirectory) throws Exception {
+  public static TestRunSummary getTestRunSummary(Path outputDirectory) throws Exception {
     return FileUtils.readOutputFileIntoJavaObject(
-        outputDirectory, TestRunner.runSummaryFileName, TestRunner.TestRunSummary.class);
+        outputDirectory, TestRunner.runSummaryFileName, TestRunSummary.class);
   }
 
   /**
@@ -594,7 +537,7 @@ public class TestRunner {
 
     // build a list of output directories that contain test run results
     List<Path> testRunOutputDirectories = new ArrayList<>();
-    TestRunner.TestRunSummary testRunSummary = null;
+    TestRunSummary testRunSummary = null;
     try {
       testRunSummary = getTestRunSummary(outputDirectory);
     } catch (Exception ex) {
@@ -662,7 +605,7 @@ public class TestRunner {
 
         // even if the test configuration didn't throw an exception, it still may have failed due to
         // a timeout
-        for (TestScriptResult.TestScriptResultSummary testScriptResultSummary :
+        for (TestScriptResultSummary testScriptResultSummary :
             runner.summary.testScriptResultSummaries) {
           if (testScriptResultSummary.isFailure) {
             testConfigFailed = true;
