+++ src/main/java/bio/terra/testrunner/runner/TestScriptResult.java
@@ -2,41 +2,12 @@ package bio.terra.testrunner.runner;
 
 import bio.terra.testrunner.common.BasicStatistics;
 import bio.terra.testrunner.runner.config.TestScriptSpecification;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.util.List;
 import org.apache.commons.math3.stat.descriptive.DescriptiveStatistics;
 
 public class TestScriptResult {
   public List<UserJourneyResult> userJourneyResults;
   public TestScriptResultSummary summary;
-
-  /**
-   * Summary statistics are pulled out into a separate inner class for easier summary reporting.
-   * This class does not include a reference to the full TestScriptSpecification or the list of
-   * UserJourneyResults.
-   */
-  @SuppressFBWarnings(
-      value = "URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD",
-      justification = "This POJO class is used for easy serialization to JSON using Jackson.")
-  public static class TestScriptResultSummary {
-    public String testScriptName;
-    public String testScriptDescription;
-
-    public BasicStatistics elapsedTimeStatistics;
-
-    public int totalRun; // total number of user journey threads submitted to the thread pool
-    public int numCompleted; // number of user journey threads that completed
-    public int numExceptionsThrown; // number of user journey threads that threw exceptions
-
-    public boolean isFailure; // numCompleted < totalRun
-
-    public TestScriptResultSummary() {} // default constructor so Jackson can deserialize
-
-    private TestScriptResultSummary(String testScriptName, String testScriptDescription) {
-      this.testScriptName = testScriptName;
-      this.testScriptDescription = testScriptDescription;
-    }
-  }
 
   public TestScriptResult(
       TestScriptSpecification testScriptSpecification, List<UserJourneyResult> userJourneyResults) {
